+++ src/main/java/com/roku/cloudgo/pojo/BankRecordExample.java
@@ -114,52 +114,52 @@ public class BankRecordExample {
             return (Criteria) this;
         }
 
-        public Criteria andUserIdEqualTo(Integer value) {
+        public Criteria andUserIdEqualTo(Long value) {
             addCriterion("user_id =", value, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdNotEqualTo(Integer value) {
+        public Criteria andUserIdNotEqualTo(Long value) {
             addCriterion("user_id <>", value, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdGreaterThan(Integer value) {
+        public Criteria andUserIdGreaterThan(Long value) {
             addCriterion("user_id >", value, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdGreaterThanOrEqualTo(Integer value) {
+        public Criteria andUserIdGreaterThanOrEqualTo(Long value) {
             addCriterion("user_id >=", value, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdLessThan(Integer value) {
+        public Criteria andUserIdLessThan(Long value) {
             addCriterion("user_id <", value, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdLessThanOrEqualTo(Integer value) {
+        public Criteria andUserIdLessThanOrEqualTo(Long value) {
             addCriterion("user_id <=", value, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdIn(List<Integer> values) {
+        public Criteria andUserIdIn(List<Long> values) {
             addCriterion("user_id in", values, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdNotIn(List<Integer> values) {
+        public Criteria andUserIdNotIn(List<Long> values) {
             addCriterion("user_id not in", values, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdBetween(Integer value1, Integer value2) {
+        public Criteria andUserIdBetween(Long value1, Long value2) {
             addCriterion("user_id between", value1, value2, "userId");
             return (Criteria) this;
         }
 
-        public Criteria andUserIdNotBetween(Integer value1, Integer value2) {
+        public Criteria andUserIdNotBetween(Long value1, Long value2) {
             addCriterion("user_id not between", value1, value2, "userId");
             return (Criteria) this;
         }
@@ -234,52 +234,52 @@ public class BankRecordExample {
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreEqualTo(Integer value) {
+        public Criteria andUserScoreEqualTo(Long value) {
             addCriterion("user_score =", value, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreNotEqualTo(Integer value) {
+        public Criteria andUserScoreNotEqualTo(Long value) {
             addCriterion("user_score <>", value, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreGreaterThan(Integer value) {
+        public Criteria andUserScoreGreaterThan(Long value) {
             addCriterion("user_score >", value, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreGreaterThanOrEqualTo(Integer value) {
+        public Criteria andUserScoreGreaterThanOrEqualTo(Long value) {
             addCriterion("user_score >=", value, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreLessThan(Integer value) {
+        public Criteria andUserScoreLessThan(Long value) {
             addCriterion("user_score <", value, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreLessThanOrEqualTo(Integer value) {
+        public Criteria andUserScoreLessThanOrEqualTo(Long value) {
             addCriterion("user_score <=", value, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreIn(List<Integer> values) {
+        public Criteria andUserScoreIn(List<Long> values) {
             addCriterion("user_score in", values, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreNotIn(List<Integer> values) {
+        public Criteria andUserScoreNotIn(List<Long> values) {
             addCriterion("user_score not in", values, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreBetween(Integer value1, Integer value2) {
+        public Criteria andUserScoreBetween(Long value1, Long value2) {
             addCriterion("user_score between", value1, value2, "userScore");
             return (Criteria) this;
         }
 
-        public Criteria andUserScoreNotBetween(Integer value1, Integer value2) {
+        public Criteria andUserScoreNotBetween(Long value1, Long value2) {
             addCriterion("user_score not between", value1, value2, "userScore");
             return (Criteria) this;
         }
