+++ src/main/java/com/roku/cloudgo/pojo/User.java
@@ -1,13 +1,13 @@
 package com.roku.cloudgo.pojo;
 
 public class User {
-    private Integer userId;
+    private Long userId;
 
     private String userName;
 
     private String userPassword;
 
-    private Integer userTelephone;
+    private Long userTelephone;
 
     private String userEmail;
 
@@ -19,11 +19,11 @@ public class User {
 
     private String userDescription;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
@@ -43,11 +43,11 @@ public class User {
         this.userPassword = userPassword == null ? null : userPassword.trim();
     }
 
-    public Integer getUserTelephone() {
+    public Long getUserTelephone() {
         return userTelephone;
     }
 
-    public void setUserTelephone(Integer userTelephone) {
+    public void setUserTelephone(Long userTelephone) {
         this.userTelephone = userTelephone;
     }
